+++ src/utils/index.ts
@@ -10,3 +10,5 @@ export {
   loadAuthFromStorage,
   removeAuthFromStorage,
 } from './storage';
+
+export * from './rankingUtils';
